Module Dns_zone

val parse : string -> ( Dns.Name_rr_map.t, [> `Msg of string ] ) Stdlib.result

parse data attempts to parse the data, given in zone file format. It either returns the content as a map, or an error.