Module Dns_server.Authentication

Authentication, stored in a Dns_trie with privileges to operations embedded in the name.

type operation = [

| `Update

| `Transfer

| `Notify

]

The type of operations, sorted by highest ot lowest privileges, an `Update may as well carry out a `Transfer.

val all_ops : operation list

type t

The type for an authenticator.

val access : ?⁠key:'a Domain_name.t -> zone:'b Domain_name.t -> operation -> bool

access op ~key ~zone checks whether key is authorised for op on zone.